requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
boto3==1.26.144
botocore==1.29.144
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-parler==2.3
django-parler-rest==2.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Faker==18.9.0
flower==1.2.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
